feat: store sizes as integer lots plus a lot size, not floats (#226)
Breaking: the on-disk schema goes 3.0 -> 4.0.

Every `volume` and `fill` column is now a whole number of lots (int64)
instead of a double in the base asset. The base-asset size is
`lots * lot_size`, where lot_size is the instrument's minimum size
increment (PipelineConfig.lot_size, default 1e-8; LOBSTER sets 1, whole
shares). This is the size half of the integer-tick decision (#155), and
it fixes a defect rather than only re-expressing the data.

A price level is a running sum of adds, cancels and fills. A float sum
does not return to exactly zero when the last order leaves, so a level
landed on residue such as 5.55e-17, stayed live, and was reported as the
best bid or ask ahead of the real one. On the bundled Bitstamp sample
that corrupted the reported best bid on 25,611 of 313,565 rows (8.2%)
and the best ask on 30,096 (9.6%) -- the spread on about one row in
eleven. Integer lots cancel exactly, so a level empties or it does not,
and those counts are now zero.

The `clip(lower=0)` that stood next to the running sum shows the residue
was already known about, but it only ever caught the negative sign. A
level emptying to -5e-17 was corrected; one emptying to +5.55e-17
survived. The sign is about even, which is why this hit a tenth of rows
rather than a fifth.

Three of the four conventions this schema already follows carry size
exactly, checked against the installed packages rather than from memory:
Databento's `size` is an integer and rejects a float outright, a
Nautilus Quantity is fixed-point with an integer `raw`, and cryptofeed
asserts Decimal. Only hftbacktest uses a float qty, and it still makes
the caller declare a lot size. Databento and Nautilus are the same two
conventions #155 cited for prices.

The change reaches every size-valued column -- depth_summary's per-bin
volumes, placed_vol and filled_vol, the book snapshot's liquidity, and
the queue's ahead_volume and remaining -- so those sums are exact too.
Three float-era workarounds went with it: the Kahan compensation behind
filled_vol, the simulator's _vol_eps exhaustion tolerance, and the
LOBSTER book replay's 1e-12 level cutoff. The Bitstamp fill matcher no
longer rounds both sides to volume_decimals to make a fill and its trade
print land in the same bucket; integers match or they do not.

Loaders convert on the way in; the plots and the round-trip and export
writers convert back, so what a user sees and what another tool reads
are unchanged. lot_size travels in each Parquet file's key-value
metadata under `ob_analytics_lot_size`, next to `ob_analytics_tick_size`,
and load_data surfaces it as df.attrs["lot_size"]. Files written at
1.0-3.0 still read, as the float-size frames they are.

Where a frame may still hold float sizes -- one built by hand, or a
pre-4.0 file -- the depth engine and the engine adapter pass it through
rather than casting, so a sub-lot size such as 0.5 is not truncated to
zero.

This closes the touch-parity check added for #224: replaying an exported
session through hftbacktest's own L3 reconstruction now agrees with
depth_summary on the best bid and ask for every row across five
synthetic seeds, and Nautilus' own book agrees too. Before this, the two
disagreed on up to 78 rows a seed. The xfail markers are removed.

Golden fingerprints move; they are re-recorded in the next commit.

+++ tests/test_datasets.py
@@ -81,8 +81,10 @@ class TestLayout:
         assert str(events["timestamp"].dtype) == "datetime64[ns, UTC]"
         assert str(events["exchange_timestamp"].dtype) == "datetime64[ns, UTC]"
         assert events["price"].dtype == "int64"
-        assert events["volume"].dtype == "float64"
-        assert events["fill"].dtype == "float64"
+        # Integer lots, the size counterpart of the integer-tick price above
+        # (issue #226).
+        assert events["volume"].dtype == "int64"
+        assert events["fill"].dtype == "int64"
         assert events["action"].dtype == "category"
         assert list(events["action"].cat.categories) == [
             "created",
